Part 4: Exercise 4.21

+++ part4/controller/blogs.js
@@ -38,6 +38,28 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    console.log(decodedToken)
+
+    if (!decodedToken) {
+        return response.status(401).json({ error: 'token invalid' })
+    }
+
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+        return response.status(204).end()
+    }
+
+    console.log('Objeto user', blog.user)
+    console.log('Token', decodedToken.id)
+
+    if (blog.user.toString() !== decodedToken.id.toString()) {
+        return response
+            .status(401)
+            .json({ error: 'only the creator can delete this blog' })
+    }
+
     await Blog.findByIdAndDelete(request.params.id)
     response.status(204).end()
 })
